Android will use a default WebViewActivity if no custom activity exists
Similar to Pull Request 7 for iOS (https://github.com/calatrava/calatrava-ios/pull/7)
this changes makes it so you don't need to create an empty Activity to
use the shell views.

The change makes the WebViewActivity concrete, and it'll try to read a
pageName from the intent that created it; the PageRegistry will now
pass the pageName to the intent whenever it creates an Activity, and
will default to using WebViewActivity.class whenever it can't find a
matching Activity.