Always recompile assets on build, fix for dependency issue
Assets weren't being recompiled when a file they depend on change, this
was most evident with changing the single_page.haml and all the other
pages wouldn't recompile.

This happens because the Rake File task will only execute if the target
file either doesn't exist or its timestamp is older than the source
file. This means the task won't execute if you change the layout
without changing each file that uses it.