update gemspec to better support Linux
Right now if you run `gem install calatrava` on Linux you'll encounter an error
trying to build native extensions for Xcodeproj, as the gem hosted on RubyGems
was built on a Mac. This change moves the platform-specific dependency
installation into a "native extension" so that it can take place at gem-install
time rather than package-build time.

For more info see this blog post:

http://www.programmersparadox.com/2012/05/21/gemspec-loading-dependent-gems-based-on-the-users-system/

Note that this approach is considered by some (including the author of the
above post) to be a hack. However, the alternative is to create platform-
specific gems, which feels even less desirable (at least to me).